+++ src/features/agency/agencyComponents/AgencyCalendar.jsx
@@ -1,7 +1,7 @@
 import Calendar from "../../../package/calendar/Calendar";
 import RoundBox from "../../../package/RoundBox";
 import useLinkUpStore from "../../../shared/store/store";
-import { useAgencyCalendar } from "../agencyServices/useAgency";
+import useAgencyCalendar from "../agencyServices/useAgencyCalendar";
 import styles from "./AgencyCalendar.module.css";
 import AgencyCalendarModal from "./AgencyCalendarModal";
