--- src/features/agency/agencyServices/agencyApi.js
@@ -1,9 +1,0 @@
-import axiosInstance from "../../../shared/services/axiosInstance";
-import useLinkUpStore from "../../../shared/store/store";
-
-export const getCompaniesArtists = async () => {
-    const response = await axiosInstance.get("/api/companies/artists");
-    const data = response.data;
-    useLinkUpStore.setState({ artistArray: data });
-    return data;
-};
